hexdns_django/requirements.txt
Django
django-grpc
dnslib
git+https://github.com/TheEnbyperor/django-keycloak-auth
sentry-sdk
psycopg2-binary
gunicorn
grpcio
protobuf
django-crispy-forms
sshpubkeys
cryptography